Refact player layout
