+++ Sources/Castor/UserInterface/Player/ControlsView.swift
@@ -4,193 +4,19 @@
 //  License information is available from the LICENSE file.
 //
 
-import CoreMedia
-import GoogleCast
 import SwiftUI
 
 struct ControlsView: View {
-    private static let shortFormatter: DateComponentsFormatter = {
-        let formatter = DateComponentsFormatter()
-        formatter.allowedUnits = [.second, .minute]
-        formatter.zeroFormattingBehavior = .pad
-        return formatter
-    }()
-
-    private static let longFormatter: DateComponentsFormatter = {
-        let formatter = DateComponentsFormatter()
-        formatter.allowedUnits = [.second, .minute, .hour]
-        formatter.zeroFormattingBehavior = .pad
-        return formatter
-    }()
-
-    @EnvironmentObject private var cast: Cast
-    @StateObject private var progressTracker = CastProgressTracker(interval: .init(value: 1, timescale: 10))
-    @ObservedObject var player: CastPlayer
-    @State private var isPlaylistPresented = false
-    let device: CastDevice?
+    let player: CastPlayer
+    @Binding var isPlaylistPresented: Bool
 
     var body: some View {
         VStack {
-            informationView()
-            Spacer()
-            visualView()
-            Spacer()
-            controls()
+            SliderView(player: player)
+            PlaybackButtons(player: player)
+            SettingsButtons(player: player, isPlaylistPresented: $isPlaylistPresented)
         }
-        .disabled(!player.isActive)
-        .bind(progressTracker, to: player)
-        .sheet(isPresented: $isPlaylistPresented) {
-            PlaylistView(player: player)
-                .presentationDetents([.medium, .large])
-        }
-    }
-
-    private static func formattedTime(_ time: CMTime, duration: CMTime) -> String? {
-        guard time.isValid, duration.isValid else { return nil }
-        if duration.seconds < 60 * 60 {
-            return shortFormatter.string(from: time.seconds)!
-        }
-        else {
-            return longFormatter.string(from: time.seconds)!
-        }
-    }
-
-    private func artworkImage() -> some View {
-        ArtworkImage(url: player.metadata?.imageUrl())
-    }
-
-    private func loadingIndicator() -> some View {
-        ProgressView()
-            .tint(.white)
-            .padding(10)
-            .background(
-                Circle()
-                    .fill(Color(white: 0, opacity: 0.4))
-            )
-            .opacity(player.isBusy ? 1 : 0)
-            .animation(.default, value: player.isBusy)
-    }
-
-    private func slider() -> some View {
-        HStack {
-            if progressTracker.isProgressAvailable {
-                Slider(
-                    progressTracker: progressTracker,
-                    label: {
-                        Text("Current position")
-                    },
-                    minimumValueLabel: {
-                        label(withText: Self.formattedTime(progressTracker.time, duration: progressTracker.timeRange.duration))
-                    },
-                    maximumValueLabel: {
-                        label(withText: Self.formattedTime(progressTracker.timeRange.duration, duration: progressTracker.timeRange.duration))
-                    }
-                )
-                if player.streamType == .live {
-                    skipToDefaultButton()
-                }
-            }
-        }
-        .frame(height: 30)
-        .padding(.vertical)
-    }
-
-    @ViewBuilder
-    private func label(withText text: String?) -> some View {
-        if let text {
-            Text(text)
-                .font(.caption)
-                .monospacedDigit()
-                .foregroundColor(.primary)
-        }
-    }
-
-    private func informationView() -> some View {
-        VStack {
-            if let title = player.metadata?.title {
-                Text(title)
-            }
-            if let device {
-                Text("Connected to \(device.name ?? "receiver")")
-                    .font(.subheadline)
-                    .foregroundStyle(.secondary)
-            }
-        }
-    }
-
-    private func visualView() -> some View {
-        ZStack {
-            artworkImage()
-            loadingIndicator()
-        }
-        .padding()
-    }
-
-    private func controls() -> some View {
-        VStack {
-            settingsButtons()
-            slider()
-            playbackButtons()
-        }
-    }
-}
-
-private extension ControlsView {
-    private func playlistButton() -> some View {
-        Button {
-            isPlaylistPresented.toggle()
-        } label: {
-            Image(systemName: "square.filled.on.square")
-        }
-        .disabled(player.items.count < 2)
-    }
-
-    private func skipBackwardButton() -> some View {
-        Button(action: player.skipBackward) {
-            Image.goBackward(withInterval: cast.configuration.backwardSkipInterval)
-        }
-        .disabled(!player.canSkipBackward())
-    }
-
-    private func skipForwardButton() -> some View {
-        Button(action: player.skipForward) {
-            Image.goForward(withInterval: cast.configuration.forwardSkipInterval)
-        }
-        .disabled(!player.canSkipForward())
-    }
-
-    private func skipToDefaultButton() -> some View {
-        Button(action: player.skipToDefault) {
-            Image(systemName: "forward.end.fill")
-        }
-        .disabled(!player.canSkipToDefault())
-    }
-
-    private func stopButton() -> some View {
-        Button {
-            player.stop()
-        } label: {
-            Image(systemName: "stop.fill")
-        }
-    }
-
-    func settingsButtons() -> some View {
-        HStack(spacing: 50) {
-            CastButton(cast: cast)
-            playlistButton()
-            SettingsMenu(player: player)
-            MuteButton(cast: cast)
-        }
-        .font(.system(size: 22))
-    }
-
-    func playbackButtons() -> some View {
-        HStack(spacing: 50) {
-            skipBackwardButton()
-            PlaybackButton(player: player)
-            stopButton()
-            skipForwardButton()
-        }
-        .font(.system(size: 44))
+        .frame(maxWidth: .infinity)
+        .background(.background)
     }
 }
